Rewrite Messages API
Move messages api actions to the message controller.
