+++ src/main/java/com/fridaymastermix/ResponseBody.java
@@ -1,0 +1,70 @@
+/*
+ *    Copyright 2019 Love Löfdahl
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ *
+ */
+
+package com.fridaymastermix;
+
+import com.fasterxml.jackson.annotation.JsonInclude;
+import org.joda.time.DateTime;
+import org.joda.time.DateTimeZone;
+
+public class ResponseBody {
+    private String timestamp;
+    private int status;
+
+    @JsonInclude(JsonInclude.Include.NON_NULL)
+    private String error;
+
+    @JsonInclude(JsonInclude.Include.NON_NULL)
+    private String message;
+
+    @JsonInclude(JsonInclude.Include.NON_NULL)
+    private String path;
+
+    public ResponseBody(int status, String message) {
+        this.status = status;
+        this.message = message;
+        this.timestamp = new DateTime().toString();
+    }
+
+    public ResponseBody(int status, String error, String message, String path) {
+        this.status = status;
+        this.message = message;
+        this.error = error;
+        this.path = path;
+        this.timestamp = DateTime.now().withZone(DateTimeZone.UTC).toString();
+    }
+
+    public String getTimestamp() {
+        return timestamp;
+    }
+
+    public int getStatus() {
+        return status;
+    }
+
+    public String getError() {
+        return error;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+
+    public String getPath() {
+        return path;
+    }
+}
